modified project structure
