+++ internal/ui/banned_message.go
@@ -1,7 +1,7 @@
 package ui
 
 import (
-	"cli-project/pkg/utils/formatting"
+	"cli-project/pkg/utils"
 	"fmt"
 )
 
@@ -10,10 +10,10 @@ func (ui *UI) ShowBannedMessage() {
 	fmt.Print("\033[H\033[2J")
 
 	// Display the banned message
-	fmt.Println(formatting.Colorize("====================================", "red", "bold"))
-	fmt.Println(formatting.Colorize("         YOU ARE BANNED FROM        ", "red", "bold"))
-	fmt.Println(formatting.Colorize("            THE PLATFORM            ", "red", "bold"))
-	fmt.Println(formatting.Colorize("====================================", "red", "bold"))
+	fmt.Println(utils.Colorize("====================================", "red", "bold"))
+	fmt.Println(utils.Colorize("         YOU ARE BANNED FROM        ", "red", "bold"))
+	fmt.Println(utils.Colorize("            THE PLATFORM            ", "red", "bold"))
+	fmt.Println(utils.Colorize("====================================", "red", "bold"))
 
 	fmt.Println("\nPress any key to go back...")
