+++ internal/ui/signup_page.go
@@ -2,9 +2,7 @@ package ui
 
 import (
 	"cli-project/internal/domain/models"
-	"cli-project/pkg/utils/data_cleaning"
-	"cli-project/pkg/utils/emojis"
-	"cli-project/pkg/utils/formatting"
+	"cli-project/pkg/utils"
 	"cli-project/pkg/validation"
 	"fmt"
 	"golang.org/x/crypto/ssh/terminal"
@@ -16,31 +14,31 @@ func (ui *UI) ShowSignupPage() {
 	// Clear the screen
 	fmt.Print("\033[H\033[2J")
 
-	fmt.Println(formatting.Colorize("====================================", "cyan", "bold"))
-	fmt.Println(formatting.Colorize("               Signup               ", "cyan", "bold"))
-	fmt.Println(formatting.Colorize("====================================", "cyan", "bold"))
+	fmt.Println(utils.Colorize("====================================", "cyan", "bold"))
+	fmt.Println(utils.Colorize("               Signup               ", "cyan", "bold"))
+	fmt.Println(utils.Colorize("====================================", "cyan", "bold"))
 
 	// Read Username
 	var username string
 	for {
-		fmt.Print(formatting.Colorize("Username: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Username: ", "yellow", ""))
 		username, _ = ui.reader.ReadString('\n')
 		username = strings.TrimSuffix(username, "\n")
-		username = data_cleaning.CleanString(username)
+		username = utils.CleanString(username)
 
 		if validation.ValidateUsername(username) {
 			unique, err := ui.authService.IsUsernameUnique(username)
 			if err != nil {
-				fmt.Println(emojis.Error, err)
+				fmt.Println(utils.ErrorEmoji, err)
 				continue
 			}
 			if !unique {
-				fmt.Println(emojis.Info, "Username already taken. Choose another username.")
+				fmt.Println(utils.InfoEmoji, "Username already taken. Choose another username.")
 				continue
 			}
 
 		} else {
-			fmt.Println(emojis.Error, "Invalid username. It should be between 4 and 20 characters long, should not be only numbers and contain no spaces.")
+			fmt.Println(utils.ErrorEmoji, "Invalid username. It should be between 4 and 20 characters long, should not be only numbers and contain no spaces.")
 			continue
 		}
 		break
@@ -49,26 +47,26 @@ func (ui *UI) ShowSignupPage() {
 	// Read Password
 	var password, confirmPassword string
 	for {
-		fmt.Print(formatting.Colorize("Password: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Password: ", "yellow", ""))
 		passwordBytes, _ := terminal.ReadPassword(int(os.Stdin.Fd()))
 		password = string(passwordBytes)
 		password = strings.TrimSpace(password)
 		fmt.Println()
 
 		// Read Confirm Password
-		fmt.Print(formatting.Colorize("Confirm Password: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Confirm Password: ", "yellow", ""))
 		confirmPasswordBytes, _ := terminal.ReadPassword(int(os.Stdin.Fd()))
 		confirmPassword = string(confirmPasswordBytes)
 		confirmPassword = strings.TrimSpace(confirmPassword)
 		fmt.Println()
 
 		if password != confirmPassword {
-			fmt.Println(emojis.Error, "Passwords do not match. Please try again.")
+			fmt.Println(utils.ErrorEmoji, "Passwords do not match. Please try again.")
 			continue
 		}
 
 		if !validation.ValidatePassword(password) {
-			fmt.Println(emojis.Error, "Invalid password. It must be at least 8 characters long and include at least 1 uppercase & lowercase letters, 1 digit, and 1 special character.")
+			fmt.Println(utils.ErrorEmoji, "Invalid password. It must be at least 8 characters long and include at least 1 uppercase & lowercase letters, 1 digit, and 1 special character.")
 			continue
 		}
 
@@ -78,12 +76,12 @@ func (ui *UI) ShowSignupPage() {
 	// Read Name
 	var name string
 	for {
-		fmt.Print(formatting.Colorize("Name: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Name: ", "yellow", ""))
 		name, _ = ui.reader.ReadString('\n')
 		name = strings.TrimSuffix(name, "\n")
 		name = strings.TrimSpace(name)
 		if !validation.ValidateName(name) {
-			fmt.Println(emojis.Error, "Invalid name. It should be 3 to 30 characters long and contain only letters and spaces.")
+			fmt.Println(utils.ErrorEmoji, "Invalid name. It should be 3 to 30 characters long and contain only letters and spaces.")
 			continue
 		}
 		break
@@ -92,26 +90,26 @@ func (ui *UI) ShowSignupPage() {
 	// Read Email
 	var email string
 	for {
-		fmt.Print(formatting.Colorize("Email: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Email: ", "yellow", ""))
 		email, _ = ui.reader.ReadString('\n')
 		email = strings.TrimSuffix(email, "\n")
-		email = data_cleaning.CleanString(email)
+		email = utils.CleanString(email)
 
 		if check1, check2 := validation.ValidateEmail(email); check1 == true && check2 == true {
 			unique, err := ui.authService.IsEmailUnique(email)
 			if err != nil {
-				fmt.Println(emojis.Error, "Error checking email uniqueness. Try again.")
+				fmt.Println(utils.ErrorEmoji, "Error checking email uniqueness. Try again.")
 				continue
 			}
 			if !unique {
-				fmt.Println(emojis.Info, "Email already registered. Use a different email.")
+				fmt.Println(utils.InfoEmoji, "Email already registered. Use a different email.")
 				continue
 			}
 		} else if check1 == false {
-			fmt.Println(emojis.Error, "Invalid email format.")
+			fmt.Println(utils.ErrorEmoji, "Invalid email format.")
 			continue
 		} else if check2 == false {
-			fmt.Println(emojis.Error, "Invalid email domain. We only support gmail, outlook, yahoo, hotmail, icloud, watchguard emails.")
+			fmt.Println(utils.ErrorEmoji, "Invalid email domain. We only support gmail, outlook, yahoo, hotmail, icloud, watchguard emails.")
 			continue
 		}
 		break
@@ -120,7 +118,7 @@ func (ui *UI) ShowSignupPage() {
 	// Read Leetcode Username
 	var LeetcodeID string
 	for {
-		fmt.Print(formatting.Colorize("Leetcode Username: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Leetcode Username: ", "yellow", ""))
 		LeetcodeID, _ = ui.reader.ReadString('\n')
 		LeetcodeID = strings.TrimSuffix(LeetcodeID, "\n")
 		LeetcodeID = strings.TrimSpace(LeetcodeID)
@@ -129,22 +127,22 @@ func (ui *UI) ShowSignupPage() {
 		isUnique, err := ui.authService.IsLeetcodeIDUnique(LeetcodeID)
 
 		if err != nil {
-			fmt.Println(emojis.Error, "Error checking Leetcode ID uniqueness. Try again.")
+			fmt.Println(utils.ErrorEmoji, "Error checking Leetcode ID uniqueness. Try again.")
 			continue
 		}
 		if !isUnique {
-			fmt.Println(emojis.Error, "Leetcode ID is already taken. Please choose a different ID.")
+			fmt.Println(utils.ErrorEmoji, "Leetcode ID is already taken. Please choose a different ID.")
 			continue
 		}
 
 		// Validate Leetcode Username with Leetcode API
 		exists, err := ui.authService.ValidateLeetcodeUsername(LeetcodeID)
 		if err != nil {
-			fmt.Println(emojis.Error, "Error validating Leetcode username:", err)
+			fmt.Println(utils.ErrorEmoji, "Error validating Leetcode username:", err)
 			continue
 		}
 		if !exists {
-			fmt.Println(emojis.Error, "Leetcode username does not exist.")
+			fmt.Println(utils.ErrorEmoji, "Leetcode username does not exist.")
 			continue
 		}
 		break
@@ -152,13 +150,13 @@ func (ui *UI) ShowSignupPage() {
 
 	var organisation string
 	for {
-		fmt.Print(formatting.Colorize("Organisation: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Organisation: ", "yellow", ""))
 		organisation, _ = ui.reader.ReadString('\n')
 		organisation = strings.TrimSuffix(organisation, "\n")
-		organisation = data_cleaning.CleanString(organisation)
+		organisation = utils.CleanString(organisation)
 		valid, err := validation.ValidateOrganizationName(organisation)
 		if !valid {
-			fmt.Println(emojis.Error, err)
+			fmt.Println(utils.ErrorEmoji, err)
 			continue
 		}
 		break
@@ -166,13 +164,13 @@ func (ui *UI) ShowSignupPage() {
 
 	var country string
 	for {
-		fmt.Print(formatting.Colorize("Country: ", "yellow", ""))
+		fmt.Print(utils.Colorize("Country: ", "yellow", ""))
 		country, _ = ui.reader.ReadString('\n')
 		country = strings.TrimSuffix(country, "\n")
-		country = data_cleaning.CleanString(country)
+		country = utils.CleanString(country)
 		valid, err := validation.ValidateCountryName(country)
 		if !valid {
-			fmt.Println(emojis.Error, err)
+			fmt.Println(utils.ErrorEmoji, err)
 			continue
 		}
 		break
@@ -194,11 +192,11 @@ func (ui *UI) ShowSignupPage() {
 	// Call Signup Service
 	err := ui.userService.Signup(&user)
 	if err != nil {
-		fmt.Println(emojis.Error, "Signup failed:", err)
+		fmt.Println(utils.ErrorEmoji, "Signup failed:", err)
 		return
 	}
 
-	fmt.Println(emojis.Success, "Signup successful!")
+	fmt.Println(utils.SuccessEmoji, "Signup successful!")
 
 	return
 }
