+++ internal/ui/update_progress_page.go
@@ -1,7 +1,7 @@
 package ui
 
 import (
-	"cli-project/pkg/utils/formatting"
+	"cli-project/pkg/utils"
 	"fmt"
 	"strings"
 )
@@ -11,18 +11,18 @@ func (ui *UI) UpdateProgressPage() {
 		// Clear the screen
 		fmt.Print("\033[H\033[2J")
 
-		fmt.Println(formatting.Colorize("====================================", "cyan", "bold"))
-		fmt.Println(formatting.Colorize("           UPDATE PROGRESS          ", "cyan", "bold"))
-		fmt.Println(formatting.Colorize("====================================", "cyan", "bold"))
-		fmt.Println(formatting.Colorize("1. Update progress", "", ""))
-		fmt.Println(formatting.Colorize("2. Go back", "", ""))
+		fmt.Println(utils.Colorize("====================================", "cyan", "bold"))
+		fmt.Println(utils.Colorize("           UPDATE PROGRESS          ", "cyan", "bold"))
+		fmt.Println(utils.Colorize("====================================", "cyan", "bold"))
+		fmt.Println(utils.Colorize("1. Update progress", "", ""))
+		fmt.Println(utils.Colorize("2. Go back", "", ""))
 
-		fmt.Print(formatting.Colorize("Enter your choice: ", "yellow", "bold"))
+		fmt.Print(utils.Colorize("Enter your choice: ", "yellow", "bold"))
 		choice, err := ui.reader.ReadString('\n')
 		choice = strings.TrimSuffix(choice, "\n")
 		choice = strings.TrimSpace(choice)
 		if err != nil {
-			fmt.Println(formatting.Colorize("Error reading input:", "red", "bold"), err)
+			fmt.Println(utils.Colorize("Error reading input:", "red", "bold"), err)
 			return
 		}
 
@@ -32,7 +32,7 @@ func (ui *UI) UpdateProgressPage() {
 		case "2":
 			return
 		default:
-			fmt.Println(formatting.Colorize("Invalid choice. Please select a valid option.", "red", "bold"))
+			fmt.Println(utils.Colorize("Invalid choice. Please select a valid option.", "red", "bold"))
 		}
 
 	}
@@ -41,14 +41,14 @@ func (ui *UI) UpdateProgressPage() {
 func (ui *UI) updateProgress() {
 
 	// Update the user's progress by marking the selected question as done
-	fmt.Println(formatting.Colorize("Fetching progress updates...", "green", ""))
+	fmt.Println(utils.Colorize("Fetching progress updates...", "green", ""))
 	err := ui.userService.UpdateUserProgress()
 
 	if err != nil {
-		fmt.Println(formatting.Colorize("Failed to update progress: ", "red", "bold"), err)
+		fmt.Println(utils.Colorize("Failed to update progress: ", "red", "bold"), err)
 		return
 	} else {
-		fmt.Println(formatting.Colorize("Updated progress successfully", "green", "bold"))
+		fmt.Println(utils.Colorize("Updated progress successfully", "green", "bold"))
 	}
 
 	fmt.Println("\nPress any key to go back...")
