--- pkg/utils/data_cleaning/clean_string.go
@@ -1,7 +1,0 @@
-package data_cleaning
-
-import "strings"
-
-func CleanString(input string) string {
-	return strings.ToLower(strings.TrimSpace(input))
-}
