--- pkg/utils/data_cleaning/clean_tags.go
@@ -1,11 +1,0 @@
-package data_cleaning
-
-import "strings"
-
-func CleanTags(tags string) []string {
-	tagList := strings.Split(tags, ",")
-	for i, tag := range tagList {
-		tagList[i] = CleanString(tag)
-	}
-	return tagList
-}
