--- pkg/utils/emojis/emojis.go
@@ -1,16 +1,0 @@
-package emojis
-
-const (
-	Signup   = "✍️"
-	Login    = "🔑"
-	Exit     = "🚪"
-	Error    = "❌"
-	Success  = "✅"
-	Profile  = "👤"
-	Stats    = "📊"
-	Settings = "⚙️"
-	Question = "❓"
-	Info     = "ℹ️"
-	Back     = "🔙"
-	View     = "👁️"
-)
