+++ pkg/utils/hashing.go
@@ -1,4 +1,4 @@
-package password
+package utils
 
 import (
 	"golang.org/x/crypto/bcrypt"
