--- pkg/validation/validate_questionID.go
@@ -1,14 +1,0 @@
-package validation
-
-import (
-	"errors"
-	"strconv"
-)
-
-func ValidateQuestionID(questionID string) (bool, error) {
-	qid, err := strconv.Atoi(questionID)
-	if err != nil || qid <= 0 {
-		return false, errors.New("invalid question ID : must be a positive number")
-	}
-	return true, nil
-}
